+++ src/layout/header.js
@@ -1,4 +1,4 @@
-import Navigation from "../components/navigation";
+import Navigation from "@/components/navigation";
 
 const Header = () => {
   return `
